+++ modules/app/codec/mod.rs
@@ -306,7 +306,7 @@ pub extern "C" fn module_step(state: *mut u8) -> i32 {
         if s.format == FMT_DETECTING {
             let sys = s.sys();
             let in_poll = (sys.channel_poll)(s.in_chan, POLL_IN);
-            if in_poll <= 0 || ((in_poll as u8) & POLL_IN) == 0 {
+            if in_poll <= 0 || ((in_poll as u32) & POLL_IN) == 0 {
                 return 0;
             }
 
@@ -360,7 +360,7 @@ pub extern "C" fn module_step(state: *mut u8) -> i32 {
         {
             let sys = s.sys();
             let in_poll = (sys.channel_poll)(s.in_chan, POLL_IN | POLL_HUP);
-            if (in_poll as u8) & POLL_HUP != 0 {
+            if (in_poll as u32) & POLL_HUP != 0 {
                 // Flush our input: clears old encoded data and the eof_flag
                 dev_channel_ioctl(sys, s.in_chan, IOCTL_FLUSH, core::ptr::null_mut());
                 dev_log(sys, 3, b"[dec] rst".as_ptr(), 9);
